+++ src/middlewear/validateApiKey/index.ts
@@ -4,15 +4,35 @@ const EARMARK_API_KEY = process.env.EARMARK_API_KEY
 
 const validateApiKey = (req:any, res:any, next:any) => {
     const reqApiKey = req.headers['earmark-api-key'];
-    if (reqApiKey !== EARMARK_API_KEY) {
+    if (!EARMARK_API_KEY) {
+        res.status(503);
+        res.json({
+            error: "Server missing earmark-api-key",
+            message: "No earmark-api-key found server-side, check enviroment variables.",
+            provided_api_key: reqApiKey,
+        });
+        res.end();
+        console.log('server-side error, earmark-api-key not found. provided key from client: ', reqApiKey);
+        return false;
+    } else if (!reqApiKey) {
+        res.status(401);
+        res.json({
+            error: "Unauthorized - Missing API key",
+            message: "Please provide an api key",
+            provided_api_key: undefined,
+        });
+        res.end();
+        console.log('missing api key')
+        return false;
+    } else if (reqApiKey !== EARMARK_API_KEY && reqApiKey) {
         res.status(401);
         res.json({
             error: "Unauthorized - invalid api key",
             message: "Please provide a valid api key",
-            api_key: reqApiKey,
+            provided_api_key: reqApiKey,
         });
         res.end();
-        console.log('invalid key')
+        console.log('invalid api key. provided key from client: ', reqApiKey);
         return false;
     }
     next();